+++ modules/const/const.py
@@ -1,0 +1,10 @@
+# ========================================================== #
+# [処理名]
+# 定数定義
+# ========================================================== #
+
+# ChromeWebdriverのパス
+CHROME_DRIVER = 'C:/tools/chromedriver.exe'
+
+# キャプチャファイルのパス
+IMG_PATH = './img/'
